tests/wpt/tests/html/canvas/element/text/2d.text.fontVariantCaps2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.fontVariantCaps2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.fontVariantCaps2</h1>
<p class="desc">Testing small caps setting in fontVariant</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing small caps setting in fontVariant");
_addTest(function(canvas, ctx) {

  ctx.font = "small-caps 32px serif";
  // "mismatch" test, to verify that small-caps does change the rendering.
  smallCaps_len = ctx.measureText("Hello World").width;

  ctx.font = "32px serif";
  normalCaps_len = ctx.measureText("Hello World").width;
  _assert(smallCaps_len != normalCaps_len, "smallCaps_len != normalCaps_len");

});
</script>
